requirements.txt
einops==0.7.0
einops_exts==0.0.4
tqdm==4.66.1
ema_pytorch==0.2.3
accelerate==0.23.0
transformers==4.34.0
pytorch_fid==0.3.0
matplotlib==3.8.0
pynvml==11.5.0
tensorboard
imageio[ffmpeg]
git+https://github.com/hassony2/torch_videovision
git+https://github.com/facebookresearch/r3m
-e ./flowdiffusion
-e ./metaworld
ai2thor==4.0.0
cython==3.0.0a10